Read Stock Synthesis starter file as a list

Read Stock Synthesis starter file as a list

Usage

SS_readstarter(file = "starter.ss", verbose = TRUE)

Arguments

file

Filename either with full path or relative to working directory.

See the formal arguments for a possible default filename.

verbose

A logical value specifying if output should be printed to the screen.

Value

A list with one element for each line of input values. List elements containing the name of the control and data file are particularly helpful, i.e., ctlfile and datfile, respectively.

See also

Author

Ian G. Taylor, Kathryn L. Doering, Kelli F. Johnson

Examples

starter_list <- SS_readstarter(
  system.file("extdata", "simple_small", "starter.ss",
    package = "r4ss"
  ),
  verbose = FALSE
)

# The following lines should be TRUE and demonstrate how you can know the
# names of the control and data file given information in the starter file.
starter_list[["ctlfile"]] == "simple_control.ss"
#> [1] FALSE
starter_list[["datfile"]] == "simple_data.ss"
#> [1] FALSE