Convert composition data to long format

Convert data frame containing length or generalized size composition data from wide to long format.

Usage

comp2long(x, measure = NULL, zero = TRUE)

Arguments

x

A data frame containing length or generalized size composition data, or a list containing such a data frame.

measure

An optional string indicating the type of measurement, e.g., "weight".

zero

Whether zero frequencies should be included in the output.

Value

Data frame containing the columns

year  month  fleet  sex  part  Nsamp  length  freq

for length composition data, or

method  year  month  fleet  sex  part  Nsamp  size  freq

for generalized size composition data.

The second to last column can be named explicitly via the measure argument.

Details

If x is not a data frame, the function will look for data frames called x[["lencomp"]] or x[["dat"]][["lencomp"]].

Note

In r4ss list objects, length composition data are stored as lencomp, while generalized size composition data are stored as sizefreq_data_list[[i]].

See also

Examples

# Read composition data from example model
path <- system.file("extdata", "simple_small", package = "r4ss")
inputs <- SS_read(path)
dat <- SS_readdat(file.path(path, "data.ss"), verbose = FALSE)
x <- dat[["lencomp"]]

# Main argument can be a list or data frame
head(comp2long(inputs))
#>   year month fleet sex part Nsamp length freq
#> 1 2011     7     1   3    0    50     26    0
#> 2 2012     7     1   3    0    50     26    0
#> 3 2013     7     1   3    0    50     26    0
#> 4 2014     7     1   3    0    50     26    0
#> 5 2015     7     1   3    0    50     26    0
#> 6 2016     7     1   3    0    50     26    1
head(comp2long(dat))
#>   year month fleet sex part Nsamp length freq
#> 1 2011     7     1   3    0    50     26    0
#> 2 2012     7     1   3    0    50     26    0
#> 3 2013     7     1   3    0    50     26    0
#> 4 2014     7     1   3    0    50     26    0
#> 5 2015     7     1   3    0    50     26    0
#> 6 2016     7     1   3    0    50     26    1
head(comp2long(x))
#>   year month fleet sex part Nsamp length freq
#> 1 2011     7     1   3    0    50     26    0
#> 2 2012     7     1   3    0    50     26    0
#> 3 2013     7     1   3    0    50     26    0
#> 4 2014     7     1   3    0    50     26    0
#> 5 2015     7     1   3    0    50     26    0
#> 6 2016     7     1   3    0    50     26    1

# Rename second to last column
head(comp2long(x, measure = "weight"))
#>   year month fleet sex part Nsamp weight freq
#> 1 2011     7     1   3    0    50     26    0
#> 2 2012     7     1   3    0    50     26    0
#> 3 2013     7     1   3    0    50     26    0
#> 4 2014     7     1   3    0    50     26    0
#> 5 2015     7     1   3    0    50     26    0
#> 6 2016     7     1   3    0    50     26    1

# Shrink output by omitting zero frequencies
nrow(comp2long(x))
#> [1] 800
nrow(comp2long(x, zero = FALSE))
#> [1] 418