Table of length and age comp Francis weights (doesn't yet include other data types)

Table of length and age comp Francis weights (doesn't yet include other data types)

Usage

table_compweight(
  replist,
  dir = NULL,
  caption = paste("Data weightings applied to compositions",
    "according to the `Francis` method. `Obs.` refers to the number of unique",
    "composition vectors included in the likelihood. `N input` and `N adj.`",
    "refer to the sample sizes of those vectors before and after being adjusted",
    "by the the weights."),
  caption_CAAL = "`CAAL` is conditional age-at-length data.",
  verbose = TRUE
)

Arguments

replist

A list object created by SS_output().

dir

Directory where the .rda files will be written. The default value is NULL where a table folder will be created where the Report.sso file is located associated with replist.

caption

A character string for the caption.

caption_CAAL

Additional text added to the caption for models with conditional age at length data.

verbose

A logical value specifying if output should be printed to the screen.

Value

A list containing the table and caption.

See also

Author

Kelli F. Johnson, Ian G. Taylor

Examples

if (FALSE) { # \dontrun{
# Load the model output
output <- r4ss::SS_output()
# Create the table
table_compweight(output)
} # }