Table of estimated time series

Adapted from Lingcod 2021: https://github.com/pfmc-assessments/lingcod/blob/main/R/table_ts.R

Usage

table_ts(
  replist,
  dir = NULL,
  caption = "Time series of population estimates for the base model.",
  verbose = TRUE
)

Arguments

replist

A list object created by SS_output().

dir

Directory where the .rda files will be written. The default value is NULL where a table folder will be created where the Report.sso file is located associated with replist.

caption

A character string for the caption.

verbose

A logical value specifying if output should be printed to the screen.

See also

Author

Kelli F. Johnson, Ian G. Taylor

Examples

if (FALSE) { # \dontrun{
table_ts(model)
} # }