Get output from multiple Stock Synthesis models.

Apply the function SS_output() multiple times and save output as individual objects or a list of lists.

Usage

SSgetoutput(
  keyvec = NULL,
  dirvec = NULL,
  getcovar = TRUE,
  getcomp = TRUE,
  forecast = TRUE,
  verbose = TRUE,
  ncols = lifecycle::deprecated(),
  listlists = TRUE,
  underscore = FALSE,
  save.lists = FALSE
)

Arguments

keyvec

A vector of strings that are appended to the output files from each model if models are all in one directory. Default=NULL.

dirvec

A vector of directories (full path or relative to working directory) in which model output is located. Default=NULL.

getcovar

Choice to read or not read covar.sso output (saves time and memory). Default=TRUE.

getcomp

Choice to read or not read CompReport.sso output (saves time and memory). Default=TRUE.

forecast

Choice to read or not read forecast quantities. Default=FALSE.

verbose

A logical value specifying if output should be printed to the screen.

ncols

Deprecated. Value is now calculated automatically.

listlists

Save output from each model as a element of a list (i.e. make a list of lists). Default = TRUE.

underscore

Add an underscore '_' between any file names and any keys in keyvec. Default=FALSE.

save.lists

Save each list of parsed output as a .Rdata file (with default filenaming convention based on iteration and date stamp.

See also

Author

Ian Taylor