Gather information for the NOAA Species Information System (SIS)

Processes model results contained in the list created by SS_output() in a format that is more convenient for submission to SIS. Currently the results are returned invisibly as a list of two tables and written to a CSV file from which results could be copied into SIS. In the future some more direct link could be explored to avoid the manual copy step.

Usage

get_SIS_info(
  model,
  dir = NULL,
  writecsv = TRUE,
  stock = "StockName",
  final_year = 2019,
  data_year = NULL,
  sciencecenter = "NWFSC",
  Mgt_Council = "NA"
)

Arguments

model

Output from SS_output

dir

A file path to the directory of interest. The default value is dir = NULL, which leads to using the current working directory.

writecsv

Write results to a CSV file (where the name will have the format "[stock]_2019_SIS_info.csv" where stock is an additional input

stock

String to prepend id info to filename for CSV file

final_year

Year of assessment and reference points (typically will be model[["endyr"]] + 1)

data_year

Last year of of timeseries data

sciencecenter

Origin of assessment report

Mgt_Council

Council jurisdiction. Currently the only option outside of the default is Gulf of Mexico ("GM")

See also

Author

Ian G. Taylor, Andi Stephens, LaTreese S. Denson

Examples

if (FALSE) {
# directory with the model output
mydir <- file.path(path.package("r4ss"), "extdata/simple_small")
# read the model output
model <- SS_output(dir = mydir)
# run get_SIS_info:
info <- get_SIS_info(model, stock = "SimpleExample")
}